Move the calling convention types used in the new calling convention modopt-based encoding down to System.Private.CoreLib/System.Runtime.
