+++ src/libraries/System.Private.CoreLib/src/System/Runtime/CompilerServices/CallingConventions.cs
@@ -1,0 +1,23 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Runtime.CompilerServices
+{
+    public class CallConvCdecl
+    {
+        public CallConvCdecl() { }
+    }
+    public class CallConvFastcall
+    {
+        public CallConvFastcall() { }
+    }
+    public class CallConvStdcall
+    {
+        public CallConvStdcall() { }
+    }
+    public class CallConvThiscall
+    {
+        public CallConvThiscall() { }
+    }
+}
