+++ src/libraries/System.Runtime.CompilerServices.VisualC/ref/System.Runtime.CompilerServices.VisualC.cs
@@ -7,22 +7,6 @@
 
 namespace System.Runtime.CompilerServices
 {
-    public partial class CallConvCdecl
-    {
-        public CallConvCdecl() { }
-    }
-    public partial class CallConvFastcall
-    {
-        public CallConvFastcall() { }
-    }
-    public partial class CallConvStdcall
-    {
-        public CallConvStdcall() { }
-    }
-    public partial class CallConvThiscall
-    {
-        public CallConvThiscall() { }
-    }
     public static partial class CompilerMarshalOverride
     {
     }
