+++ src/libraries/System.Runtime.CompilerServices.VisualC/ref/System.Runtime.CompilerServices.VisualC.csproj
@@ -6,6 +6,7 @@
   <ItemGroup>
     <Compile Include="System.Runtime.CompilerServices.VisualC.cs" />
     <Compile Include="System.Runtime.CompilerServices.VisualC.manual.cs" />
+    <Compile Include="System.Runtime.CompilerServices.VisualC.TypeForwards.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\System.Runtime\ref\System.Runtime.csproj" />
