+++ src/libraries/System.Runtime/ref/System.Runtime.cs
@@ -8856,6 +8856,22 @@ namespace System.Runtime.CompilerServices
         public void SetStateMachine(System.Runtime.CompilerServices.IAsyncStateMachine stateMachine) { }
         public void Start<TStateMachine>(ref TStateMachine stateMachine) where TStateMachine : System.Runtime.CompilerServices.IAsyncStateMachine { }
     }
+    public partial class CallConvCdecl
+    {
+        public CallConvCdecl() { }
+    }
+    public partial class CallConvFastcall
+    {
+        public CallConvFastcall() { }
+    }
+    public partial class CallConvStdcall
+    {
+        public CallConvStdcall() { }
+    }
+    public partial class CallConvThiscall
+    {
+        public CallConvThiscall() { }
+    }
     [System.AttributeUsageAttribute(System.AttributeTargets.Parameter, AllowMultiple=false, Inherited=false)]
     public sealed partial class CallerArgumentExpressionAttribute : System.Attribute
     {
